Added room support

+++ app/src/main/java/com/tobibur/subalarm/data/DummyAlarms.kt
@@ -1,20 +1,22 @@
 package com.tobibur.subalarm.data
 
-import kotlin.time.Clock
+import com.tobibur.subalarm.domain.model.Alarm
+import com.tobibur.subalarm.domain.model.SubAlarm
 
 object DummyAlarms {
 
     val alarms = mutableListOf<Alarm>()
+
     init {
-        for(i in 0..10){
+        for (i in 0..10) {
             alarms.add(
                 Alarm(
-                    id = i,
+                    id = i.toLong(),
                     title = "Alarm no $i",
                     time = System.currentTimeMillis(),
                     subAlarms = listOf(
                         SubAlarm(
-                            id = 1,
+                            id = 1L,
                             title = "Sub Alarm 1",
                             time = System.currentTimeMillis().plus(600_000)
                         )
